requirements.txt
Flask==2.3.3
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
flask-swagger-ui==4.11.1
apispec==6.3.0
marshmallow==3.20.2
PyYAML==6.0.2
